Add @schema/validation annotation
Following discussions with
[gcheadle-vmware](https://github.com/vmware-tanzu/carvel-ytt/issues/604#issuecomment-1034353974)
and @cppforlife, it became clear that @assert/validate ought to always
be consumed in a step after overlays are applied for the current
pipeline.

When writing schema, the author intends not for the schema document
_itself_ to be validated, but to validate the _produced_ document (of
which the schema is a basis for).

Authors need a way to declare the intention for a validation to be
attached to the type composite: @schema/validation.