replace confusing wording
The word `ensure` in this sentence makes the expected behavior ambiguous:

```
If the resource is not present already, then kapp uses the `exists` operation and ensures that the resource exists in Kubernetes. 
```

Using `ensures` in this sentence gives the impression that it will make sure the resource will exist.
The controller does not do this, it asserts if the resource exists, but does not create it if it does not.